+++ app/views/records/show.html.erb
@@ -59,7 +59,7 @@
     <% end %>
   <% end %>
 </div>
-  
+
   <!-- TODO: Add in Clinical Data -->
   <div class="col-md-6" style="overflow-y: scroll; height:100vh;">
     <div class="btn-group" align="right">
@@ -84,6 +84,6 @@
       </ul>
     </div>
     <% patient_resource = @record.patient.entry.find { |e| e.resource.resourceType == 'Patient' }.resource %>
-    <%= react_component("HelloWorld", { data: patient_resource, patient: patient_resource, errors: patient_resource.validate }) %>
+    <%= react_component("PatientView", { data: patient_resource, patient: patient_resource, entry: @relevant_entries, errors: patient_resource.validate }) %>
   </div>
 </div>
